Add missing `Skyr_DISABLE_LIBCXX` option deceleration in top level CMakeLists.txt
The option deceleration is missing, yet it is used for setting up the
`stdlib=stdc++` flag for clang. Cmake option has the default value of `OFF`;
the value is negated in the if statement so by default the flag is appended
regardless of whether the machine has `stdc++` installed. Since some Clang
installations do not come with `stdc++`, the project cannot be built.

The default value is set to `ON` since LLVM's implementation of C++17 is not
yet complete. More can be found in the links below:
- [cppreference](https://en.cppreference.com/w/cpp/compiler_support)
- [libc++ on 17](https://libcxx.llvm.org/cxx1z_status.html)
- [libc++ homepage](https://libcxx.llvm.org/index.html)